Refactors HTTP server functions out of app.js
